pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "groq>=1.1.2",
    "jinja2>=3.1.6",
    "langchain-openai>=1.1.12",
    "logfire[fastapi]>=4.31.0",
    "nest-asyncio>=1.6.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-ai>=1.78.0",
    "pydantic-ai-slim[tavily]>=1.78.0",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.24",
    "uvicorn>=0.44.0",
]
